+++ test/Default/MinimizeCond/Input/Test.cc
@@ -13,6 +13,7 @@
 #endif
 #ifdef COND3
 // Do something
-#else
+#ifdef NESTED
 // Do something else
+#endif
 #endif