+++ test/Transitional/MinimizeCond/CorrectOutput/Test.cc
@@ -39,6 +39,7 @@ export module Test;
 #endif
 #ifdef COND3
 // Do something
-#else
+#ifdef NESTED
 // Do something else
+#endif
 #endif